(maint) Fix puppetdb package names on FreeBSD
FreeBSD ships with all current versions of puppet and make packages
available with the major version as part of the package name, e.g.
puppet6, puppetserver7, puppetdb6, puppetdb-terminus7.

Gather the major frem thu user specified version if available, or
default to the latest version if not specified to build actual package
names.